Fix ill formed variadic argument
Turns out the variadic argument expects the
argument to be of integral type. Failing to do so
leads to implementation defined behavior as
outlined here:

https://stackoverflow.com/a/50848200/4938616

The only portable solution for user-defined types
is to implement variadic templates instead. I will
leave that for later since I do not really have a
current need for disjoint ranges.